+++ server/World.cpp
@@ -45,15 +45,15 @@ void World::init(IniReaderInterface* ir_intf)
 		race8 = false;
 	else
 		race8 = true;
-	setOffset(OT_hour, ir_intf->readIntegerEntry("WorldInfo Offsets", "WorldHourOffset"), "Hour");
+	setOffset(OT_hour, (UINT)ir_intf->readIntegerEntry("WorldInfo Offsets", "WorldHourOffset"), "Hour");
 
-	setOffset(OT_minute, ir_intf->readIntegerEntry("WorldInfo Offsets", "WorldMinuteOffset"), "Minute");
+	setOffset(OT_minute, (UINT)ir_intf->readIntegerEntry("WorldInfo Offsets", "WorldMinuteOffset"), "Minute");
 
-	setOffset(OT_day, ir_intf->readIntegerEntry("WorldInfo Offsets", "WorldDayOffset"), "Day");
+	setOffset(OT_day, (UINT)ir_intf->readIntegerEntry("WorldInfo Offsets", "WorldDayOffset"), "Day");
 
-	setOffset(OT_month, ir_intf->readIntegerEntry("WorldInfo Offsets", "WorldMonthOffset"), "Month");
+	setOffset(OT_month, (UINT)ir_intf->readIntegerEntry("WorldInfo Offsets", "WorldMonthOffset"), "Month");
 
-	setOffset(OT_year, ir_intf->readIntegerEntry("WorldInfo Offsets", "WorldYearOffset"), "Year");
+	setOffset(OT_year, (UINT)ir_intf->readIntegerEntry("WorldInfo Offsets", "WorldYearOffset"), "Year");
